Last changes in pull request
