+++ FilmsApp/Presenter/DetailPresenter.swift
@@ -1,18 +1,18 @@
 import Foundation
 
 protocol DetailPresenterDelegate: AnyObject {
-    func updateData(data: DetailFilm?)
+    func updateData(data: FilmDetails?)
 }
 
 protocol DetailPresenterProtocol: AnyObject {
-    var filmDescription: DetailFilm? { get set }
+    var filmDescription: FilmDetails? { get set }
     var delegate: DetailPresenterDelegate? { get set }
     
     func getId(id: String)
 }
 
 public class DetailPresenter: DetailPresenterProtocol {
-    var filmDescription: DetailFilm?
+    var filmDescription: FilmDetails?
     var filmId: String
     
     weak var delegate: DetailPresenterDelegate?
@@ -30,7 +30,7 @@ public class DetailPresenter: DetailPresenterProtocol {
         dataTransferService.request(with: endpoints.movie(with: id)) { [weak self] result in
             switch result {
             case .success(let movieResponse):
-                let movie = DetailFilm(id: movieResponse.id ?? "", originalTitle: movieResponse.title ?? "", genres: movieResponse.genres ?? "", countries: movieResponse.countries ?? "", rating: movieResponse.imDbRating ?? "", image: movieResponse.image ?? "")
+                let movie = FilmDetails(id: movieResponse.id ?? "", originalTitle: movieResponse.title ?? "", genres: movieResponse.genres ?? "", countries: movieResponse.countries ?? "", rating: movieResponse.imDbRating ?? "", image: movieResponse.image ?? "")
                 self?.filmDescription = movie
                 self?.delegate?.updateData(data: self?.filmDescription)
