+++ FilmsApp/View/GeneralController/MainViewController.swift
@@ -1,32 +1,32 @@
 import UIKit
 
-class CustomView: UIView {
-    var labelText: String
-    init(frame: CGRect, labelText: String) {
-        self.labelText = labelText
-        super.init(frame: frame)
-        createSubViews()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func createSubViews() {
-        backgroundColor = .white
-        let label = UILabel()
-        label.textColor = UIColor.black
-        label.text = labelText
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.numberOfLines = 0
-        label.font = UIFont.systemFont(ofSize: 20, weight: UIFont.Weight.medium)
-        addSubview(label)
-        NSLayoutConstraint.activate([
-            label.centerXAnchor.constraint(equalTo: centerXAnchor),
-            label.centerYAnchor.constraint(equalTo: centerYAnchor)
-        ])
-    }
-}
+//class CustomView: UIView {
+//    var labelText: String
+//    init(frame: CGRect, labelText: String) {
+//        self.labelText = labelText
+//        super.init(frame: frame)
+//        createSubViews()
+//    }
+//    
+//    required init?(coder: NSCoder) {
+//        fatalError("init(coder:) has not been implemented")
+//    }
+//    
+//    private func createSubViews() {
+//        backgroundColor = .white
+//        let label = UILabel()
+//        label.textColor = UIColor.black
+//        label.text = labelText
+//        label.translatesAutoresizingMaskIntoConstraints = false
+//        label.numberOfLines = 0
+//        label.font = UIFont.systemFont(ofSize: 20, weight: UIFont.Weight.medium)
+//        addSubview(label)
+//        NSLayoutConstraint.activate([
+//            label.centerXAnchor.constraint(equalTo: centerXAnchor),
+//            label.centerYAnchor.constraint(equalTo: centerYAnchor)
+//        ])
+//    }
+//}
 
 class MainViewController: UIViewController {
     
@@ -57,7 +57,7 @@ class MainViewController: UIViewController {
     }
     
     private func createTableView() {
-        self.view.addSubview(moviesCollectionsTableView)
+        view.addSubview(moviesCollectionsTableView)
         NSLayoutConstraint.activate([
             moviesCollectionsTableView.topAnchor.constraint(equalTo: self.view.topAnchor, constant: 0),
             moviesCollectionsTableView.leadingAnchor.constraint(equalTo: self.view.leadingAnchor, constant: 0),
