+++ FilmsApp/View/ListController/ListViewController.swift
@@ -8,9 +8,7 @@ class ListViewController: UIViewController {
         return table
     }()
     
-    var data: FilmsCollection
     let cellReuseIndentifier = "cell"
-    let context = CoreDataManager.instance.persistentContainer.viewContext
     var presenter: ListPresenterProtocol
     
     override func viewDidLoad() {
@@ -21,7 +19,6 @@ class ListViewController: UIViewController {
     
     init(data: FilmsCollection, presenter: ListPresenterProtocol ) {
         self.presenter = presenter
-        self.data = data
         super.init(nibName: nil, bundle: nil)
         self.presenter.delegate = self
         self.presenter.loadData(data: data)
@@ -34,13 +31,7 @@ class ListViewController: UIViewController {
     func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         let deleteAction = UIContextualAction(style: .normal, title: "Favourite") { (_, _, completionHandler) in
             DispatchQueue.global().async {
-                let film = MOFilm(context: self.context)
-                if let data = try? Data(contentsOf: (self.data.films[indexPath.row].imageURL) as! URL) {
-                    film.filmImage = data
-                }
-                film.filmTitle = self.data.films[indexPath.row].title
-                //create Presenter
-                CoreDataManager.instance.saveContext()
+                self.presenter.saveFilms(title: self.presenter.titleAtindex(index: indexPath.row), image: self.presenter.imageAtindex(index: indexPath.row))
                 DispatchQueue.main.async {
                     completionHandler(true)
                 }
@@ -71,7 +62,7 @@ class ListViewController: UIViewController {
 
 extension ListViewController: UITableViewDelegate, UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return data.films.count
+        return presenter.numberOffilms()
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
@@ -80,7 +71,6 @@ extension ListViewController: UITableViewDelegate, UITableViewDataSource {
         if let customCell = cell as? MovieListCell {
             DispatchQueue.main.async {
                 customCell.updateAppearanceFor(title: self.presenter.titleAtindex(index: indexPath.row), image: self.presenter.imageAtindex(index: indexPath.row))
-                //customCell.updateAppearanceFor(content: self.data.films[indexPath.row], image: .none)
             }
             return customCell
         }
@@ -89,25 +79,15 @@ extension ListViewController: UITableViewDelegate, UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
         guard let cell = cell as? MovieListCell else { return }
-            cell.updateAppearanceFor(title: self.presenter.titleAtindex(index: indexPath.row), image: self.presenter.imageAtindex(index: indexPath.row))
-        
-        
-//        guard let url: URL? = data.films[indexPath.row].imageURL else {return}
-//        DispatchQueue.main.async { [weak self] in
-//            if let data = try? Data(contentsOf: url as! URL) , let image = UIImage(data: data)  {
-//                DispatchQueue.main.async {
-//                    cell.updateAppearanceFor(content: self?.data.films[indexPath.row], image: image)
-//                }
-//            }
-//        }
+        cell.updateAppearanceFor(title: self.presenter.titleAtindex(index: indexPath.row), image: self.presenter.imageAtindex(index: indexPath.row))
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         listTableView.deselectRow(at: indexPath, animated: true )
-        let filmId = data.films[indexPath.row].filmId
+        let filmId = self.presenter.idAtindex(index: indexPath.row)//data.films[indexPath.row].filmId
         let dataTransferService = DefaultDataTransferService(config: NetworkConfig(server:  Server(scheme: .https, host: "imdb-api.com")))
         let endpoints = DefaultMoviesEnpdoints()
-        let newViewController = DetailController(id: filmId, presenter: DetailPresenter(dataTransferService: dataTransferService, endpoints: endpoints, id: filmId))
+        let newViewController = DetailController(id: filmId ?? "", presenter: DetailPresenter(dataTransferService: dataTransferService, endpoints: endpoints, id: filmId ?? ""))
         self.navigationController?.pushViewController(newViewController, animated: true)
     }
